move scripts into bulk directory
